feat: emit classes in separated files
