--- lib/models/cycles.dart
@@ -1,6 +1,0 @@
-class PaymentCycle {
-  static const ONCE_DAY = '日';
-  static const ONCE_WEEK = '週';
-  static const ONCE_MONTH = '月';
-  static const ONCE_YEAR = '年';
-}
